+++ src/main/java/com/example/hmspfa/services/implementations/PatientServiceImpl.java
@@ -1,6 +1,7 @@
 package com.example.hmspfa.services.implementations;
 
 import com.example.hmspfa.entities.Patient;
+import com.example.hmspfa.exceptions.PatientNotFoundException;
 import com.example.hmspfa.repositories.PatientRepository;
 import com.example.hmspfa.services.PatientService;
 import jakarta.transaction.Transactional;
@@ -9,6 +10,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -23,12 +25,19 @@ public class PatientServiceImpl implements PatientService {
     }
 
     @Override
-    public Patient getPatientById(Long id) {
-        return patientRepository.findById(id).orElse(null);
+    public Patient getPatientById(Long id) throws PatientNotFoundException {
+        return patientRepository.findById(id)
+                .orElseThrow(() -> new PatientNotFoundException("Patient Not Found"));
     }
 
     @Override
-    public void deletePatient(Long id) {
+    public void deletePatient(Long id) throws PatientNotFoundException {
+        Optional<Patient> patientOptional = patientRepository.findById(id);
+
+        if (patientOptional.isEmpty()) {
+            throw new PatientNotFoundException("Patient Not Found");
+        }
+
         patientRepository.deleteById(id);
     }
